+++ scifit/tests/test_kinetics.py
@@ -199,6 +199,21 @@ class GenericKineticTest:
         )
         plt.close(axe.figure)
 
+    def test_plot_quotient_rates(self):
+        name = self.__class__.__name__
+        self.solver.fit(t=self.t)
+        axe = self.solver.plot_quotient_rates()
+        axe.figure.savefig(
+            "{}/{}_quotient_rates.{}".format(self.media_path, name, self.format)
+        )
+        plt.close(axe.figure)
+
+    def test_process_report(self):
+        name = self.__class__.__name__
+        file = r"{}_report".format(name)
+        self.solver.fit(self.t)
+        self.solver.report(file=file, path=self.media_path, mode="pdf")
+
 
 resolution = 5000